CPAN::Meta v1 "gpl" license is explicitly GPL v2
In version 1.0 through 1.4, the license "gpl" refers the GPL v2 license.
It should not be used for code licensed as GPL v1 or GPL v3. There is a
fixed list of allowed license values for meta v1, so the only fitting
option for GPLv1 or GPLv3 is "open_source", referring to any OSI
approved license.